refactor(web): unify route + component error boundaries (LUM-1954)
Three lazy-chunk failure handling primitives had diverged across the
recent bundle work and needed consolidation:

  - LUM-1939's `LazyBoundary` (component-level: Sentry.ErrorBoundary)
  - LUM-1954's `LazyRouteErrorBoundary` (route-level: useRouteError + manual capture)
  - `RootErrorBoundary` (route-level full-page UI)

Three pieces of overlapping logic, no shared "is this a chunk error?"
helper, and a real bug: `LazyRouteErrorBoundary` called
`Sentry.captureException` *and* `RouterProvider.onError` in main.tsx
captured the same event again — every router error landed in Sentry
twice with different tags.

This PR collapses them into one canonical shape:

  - `src/lib/chunk-errors.ts` exposes `isChunkLoadError(err)`. One
    predicate for chunk-fetch failures, used by every boundary.
  - `src/components/route-error-boundary.tsx` replaces both
    `RootErrorBoundary` and `LazyRouteErrorBoundary`. It uses
    `useRouteError()` and picks between two variants:
      * chunk-fail → inline "section couldn't load" + Reload
      * anything else → full-page "Something went wrong" + Reload
    Both `root-error-boundary.tsx` and `lazy-route-error-boundary.tsx`
    are deleted; routes.tsx now references the unified component at
    every layer.
  - `main.tsx` is the single Sentry capture point for router errors.
    It tags each event with `boundary: "lazy-route"` (chunk fail) or
    `boundary: "route-render"` (other). The boundary intentionally
    does NOT capture again — single source of truth, no duplicates.
  - `LazyBoundary` (component-level) tags via Sentry's `beforeCapture`
    callback with `"lazy-component"` / `"component-render"`, symmetric
    with the route tagging. Its default inline copy now matches the
    route-level chunk-fail copy so messaging is consistent across the
    app.

`CONVENTIONS.md` updated to describe the unified shape.

No behavior change vs. the prior commit on this branch beyond:
  * Sentry capture count drops from 2 → 1 per router error
  * Full-page and inline copy is now consistent end-to-end
