+++ apps/web/src/components/lazy-boundary.tsx
@@ -1,6 +1,8 @@
 import { Suspense, type ReactElement, type ReactNode } from "react";
 import * as Sentry from "@sentry/react";
 
+import { isChunkLoadError } from "@/lib/chunk-errors";
+
 interface LazyBoundaryProps {
   children: ReactNode;
   /** Rendered while the lazy chunk is loading. */
@@ -21,11 +23,20 @@ interface LazyBoundaryProps {
 }
 
 /**
- * Suspense + ErrorBoundary pair for `React.lazy` components. Catches chunk
- * load failures (which otherwise hang on Suspense forever or escape to the
- * route's `ErrorBoundary` and nuke the whole route) and reports them to
- * Sentry. Use this anywhere `React.lazy` is rendered outside a route
- * boundary.
+ * Suspense + ErrorBoundary pair for `React.lazy` components rendered
+ * *outside* a route boundary (modals, inline lazy widgets, etc.). For
+ * lazy *routes*, React Router's `ErrorBoundary` (see
+ * `RouteErrorBoundary`) catches the failure instead.
+ *
+ * Sentry captures every caught error with `tags.boundary` set to
+ * `"lazy-component"` (chunk-fetch failures) or `"component-render"`
+ * (genuine render bugs) so they're sliceable separately and
+ * symmetric with the `"lazy-route"` / `"route-render"` tagging used
+ * by `RouterProvider.onError` for route-level errors.
+ *
+ * The default error UI matches the inline copy used by
+ * `RouteErrorBoundary`'s chunk-fail variant so messaging stays
+ * consistent across the app.
  */
 export function LazyBoundary({
   children,
@@ -34,13 +45,19 @@ export function LazyBoundary({
 }: LazyBoundaryProps) {
   return (
     <Sentry.ErrorBoundary
+      beforeCapture={(scope, error) => {
+        scope.setTag(
+          "boundary",
+          isChunkLoadError(error) ? "lazy-component" : "component-render",
+        );
+      }}
       fallback={
         errorFallback ?? (
           <div
             role="alert"
             className="p-3 text-body-small-default text-[var(--content-tertiary)]"
           >
-            Failed to load this view. Please reload the page.
+            This section couldn&apos;t load. Reload the page to try again.
           </div>
         )
       }
